Update index.js with api forecast and display
Create new function the goal is to get the forecast data which is going to make an API call and then display the forecast with the API results.

1. Create a function:   function getForecast(city);

2. Create a variable for apiKey & apiUrl (must include forecast not current in the url)

3. Call function:    getForecast(“Brooklyn”);

Now how do we get the data inside the code

4. Use axios:    axios.(apiUrl).then(displayForecast);

5. Under the function displayForecast(response){
    console.log(response.data
}

To see the data of the city that was called.

No longer have to call the getForecast(“Brooklyn”) at the end of the javascript so it can be deleted.

It was replaced with adding the getForecast(response.data.city) to get entered city forecast data from the API