Update forecast & icons with API integration
Analyze the response that we are getting from the API
Connect the forecastHtml code to the response.data ie, apiUrl
The API data.daily gives a seven day prediction which also gives a timestamp which we cause to extract the day

1. First display the temperatures: Get the temperatures. Use the console.log(response.data) to see where the daily data is going to come from

2. Remove the variable days = [“Tue”, “Wed”, etc.] array.
	a. Instead of looping through these days we can loop through: response.data.daily.forEach(function(day))
		I. The day is the array of objects 0, 1, 2, 3, 4, 5, 6. Which is representative of a day. 0 = Sunday 6 = Saturday
￼

3. Replace the default max/min temperatures with the api data and parse it:
	${day.temperature.maximum}
	${day.temperature.minimum}

	a. Make sure to round the number to make it whole:
		${Math.round(day.temperature.maximum)}		${Math.round(day.temperature.minimum)}

So, what we are doing here is looping through each daily object (which is the day) to get accurate/current temperature data

Now, we are going to get the icon/emoji to display based on the condition.

4. Remove the default icon/emoji and replace it with: <img src = “${day.condition.icon_url}” />

5. Now adjusting the image size
 Best to do this is to move the class inside of the img and remove the <div></div>to target it better so only this remains:<img src = “${day.condition.icon_url}” class=“weather-forecast-icon”/>

6. in the .css file using the class=“weather-forecast-icon” selector adjust the parameters to get desired look

So now, we have the predicted forecast for future days. However all of the days/objects in the data.daily are shown. How do you get it down to desired amount of days to display? Ex: 5 days instead of 7.

7. The function: response.data.daily.forEach(function (day) {} gives us the day, but we can also add “index” to the parameter. Also add a condition to compliment the index.

response.data.daily.forEach(function (day, index) {
      If (index < 5) {
**Forecast code for current data**
   }
}

This allows for control over the array of the objects to specify a certain amount of loops before stopping. We want display 5 days of temperatures:

First loop = index 0, display data
Second loop = index 1 display data
Third loop = index 2 display data
Fourth loop = index 3 display data
Fifth loop = index 4 display data

The index value gets incremented, every time the loop is going through. So we are going through each day here. So in this case, If the index is lower than 5, run the “Forecast code for current data”, if not, don’t do it.

Now, we need to work on the day name because as of right now, it is set on default Tue or ${day}.

8. Remove Tue or ${day}.

9. Create a new function of the API timestamp that is given:

	function formatDay(timestamp) {
		date variable (make sure to multiple by 1000 because timestamps are given in milliseconds)
		days variable

		return the days (remember it is going to be a value between 0 and 6 & the days value/name will be based on the index)
}

function formatDay(timestamp) {
  let date = new Date(timestamp * 1000);
  let days = ["Sun", "Mon", "Tue", "Wed", "Thu", "Fri", "Sat"];

  return days[date.getDay()];
}

Example, if the day today is Tuesday, the index is gong to be 2, which is the third vale of the array.

9. Use the function created to place inside the **Forecast code** for the class=“weather-forecast-date”:

${formatDay(day.time)}